% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_variable.R
\name{add_continuous_variable}
\alias{add_continuous_variable}
\title{Add continuous variable}
\usage{
add_continuous_variable(
  data,
  obs_path,
  var_name,
  med_code_list,
  start_col,
  start_offset_days = 0,
  end_col,
  end_offset_days = 0,
  unique_col = "pregid",
  value_col = "value",
  exclude_previous_records = FALSE,
  previous_start_col = "pregstart",
  previous_end_col = "pregend"
)
}
\arguments{
\item{data}{Main data.table}

\item{obs_path}{Path to observation data file}

\item{var_name}{Name of the variable to add}

\item{med_code_list}{List of medical codes to extract}

\item{start_col}{Start date column name}

\item{start_offset_days}{Days to offset start (default: 0)}

\item{end_col}{End date column name}

\item{end_offset_days}{Days to offset end (default: 0)}

\item{unique_col}{Unique identifier column (default: "pregid")}

\item{value_col}{Name of the value column in event data (default: "value")}

\item{exclude_previous_records}{Whether to exclude events that fall within previous records (default: FALSE)}

\item{previous_start_col}{Start column for previous records (default: "pregstart")}

\item{previous_end_col}{End column for previous records (default: "pregend")}
}
\value{
data.table with added continuous variable
}
\description{
Add continuous variable
}
