% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_variable.R
\name{add_previous_outcome}
\alias{add_previous_outcome}
\title{Add previous outcome indicator}
\usage{
add_previous_outcome(data, outcome_col, new_col_name, time_col = "pregstart")
}
\arguments{
\item{outcome_col}{Column indicating outcome (0/1 or TRUE/FALSE)}

\item{new_col_name}{Name for new column}

\item{time_col}{Time column for ordering (default: "pregstart")}

\item{df}{data.table}
}
\value{
data.table with added previous outcome column
}
\description{
Add previous outcome indicator
}
