% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_event_and_date.R
\name{extract_event_and_date.aurum_obs}
\alias{extract_event_and_date.aurum_obs}
\title{Extract events and corresponding dates from CPRD Aurum observation table}
\usage{
\method{extract_event_and_date}{aurum_obs}(X, med_code_list)
}
\arguments{
\item{X}{A list containing the path to the CPRD Aurum observation table.}

\item{med_code_list}{A vector of medical code IDs to filter the observation records.}
}
\value{
A data.table containing the patient ID, medical code ID, event date, and value for the specified medical codes.
}
\description{
Extract events and corresponding dates from CPRD Aurum observation table
}
