% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_event_and_date.R
\name{extract_event_and_date.hes_episodes}
\alias{extract_event_and_date.hes_episodes}
\title{Extract events and corresponding dates from HES episodes table}
\usage{
\method{extract_event_and_date}{hes_episodes}(X, icd_code_list)
}
\arguments{
\item{X}{A data.table or a list containing the path to the HES episodes table.}

\item{icd_code_list}{A vector of ICD codes to filter the episodes records.}
}
\value{
A data.table containing the patient ID and the corresponding event date for the specified ICD codes.
}
\description{
Extract events and corresponding dates from HES episodes table
}
