% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_table.R
\name{extract_from_table}
\alias{extract_from_table}
\title{Extract subset of data by patient IDs}
\usage{
extract_from_table(
  table_name,
  table_path,
  id_list,
  output_path,
  select_cols = NULL
)
}
\arguments{
\item{table_name}{The name of the table to extract from (e.g., "Observation").}

\item{table_path}{The path to the Parquet file containing the table data.}

\item{id_list}{A vector of patient IDs to extract.}

\item{output_path}{The path where the extracted subset of data will be saved in Parquet format.}
}
\value{
None. The function writes the extracted data to the specified output path.
}
\description{
Extract subset of data by patient IDs
}
