% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_medcode.R
\name{map_to_medcode.read}
\alias{map_to_medcode.read}
\title{Maps READ codes to CPRD medcode IDs using a provided code dictionary.}
\usage{
\method{map_to_medcode}{read}(X, code_dict_path)
}
\arguments{
\item{X}{A vector of READ codes (either original or cleansed).}

\item{code_dict_path}{The path to the code dictionary file.}
}
\value{
A vector of CPRD medcode IDs corresponding to the input READ codes.
}
\description{
Maps READ codes to CPRD medcode IDs using a provided code dictionary.
}
