% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_medcode.R
\name{map_to_medcode.snomed}
\alias{map_to_medcode.snomed}
\title{This function maps SNOMED CT concept IDs to CPRD medcode IDs using a provided code dictionary.}
\usage{
\method{map_to_medcode}{snomed}(X, code_dict_path)
}
\arguments{
\item{X}{A vector of SNOMED CT concept IDs.}

\item{code_dict_path}{The path to the code dictionary file.}
}
\value{
A vector of CPRD medcode IDs corresponding to the input SNOMED CT concept IDs.
}
\description{
This function maps SNOMED CT concept IDs to CPRD medcode IDs using a provided code dictionary.
}
