% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_variable.R
\name{add_binary_variable}
\alias{add_binary_variable}
\title{Add binary variable (based on continuous variable logic with optional last event date)}
\usage{
add_binary_variable(data, obs_path, med_code_list, hes_epi, icd_code_list, ...)
}
\description{
Add binary variable (based on continuous variable logic with optional last event date)
}
