% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_variable.R
\name{add_continuous_variable}
\alias{add_continuous_variable}
\title{Add continuous variable (with optional last event date)}
\usage{
add_continuous_variable(
  data,
  obs_path,
  med_code_list,
  exclude_previous_records = FALSE,
  hes_epi,
  icd_code_list,
  value_col = "value",
  ...
)
}
\description{
Add continuous variable (with optional last event date)
}
