% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_variable.R
\name{add_previous_condition}
\alias{add_previous_condition}
\title{Add previous condition indicator
, to indicate whether a patient had the condition in the previous time period (e.g., gestational diabetes in the previous pregnancy)}
\usage{
add_previous_condition(data, outcome_col, new_col_name = NULL, time_col = NULL)
}
\description{
Add previous condition indicator
, to indicate whether a patient had the condition in the previous time period (e.g., gestational diabetes in the previous pregnancy)
}
