% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_aurum.R
\name{extract_from_table}
\alias{extract_from_table}
\title{Extract subset of data by specified column and values}
\usage{
extract_from_table(
  table_name,
  table_path,
  col_name,
  value_list,
  output_path,
  select_cols = NULL
)
}
\arguments{
\item{table_name}{The name of the table to extract from (e.g., "Observation").}

\item{table_path}{The path to the Parquet file containing the table data.}

\item{col_name}{The name of the column to filter on (e.g., "patid").}

\item{value_list}{A vector of values to filter the specified column by.}

\item{output_path}{The path where the extracted subset of data will be saved in Parquet format.}

\item{select_cols}{Optional vector of column names to select from the table. If NULL, all columns will be selected.}
}
\value{
A data.table containing the extracted subset of data.
}
\description{
Extract subset of data by specified column and values
}
